Minor code style corrections
Introducing of some "closing" constant to enumerator, defining the total number of mutually exclusive modes/states/etc... instead of separate constant is more convenient and is used commonly.
And we have to avoid any warnings in the code. So, reminder of % operator should be explicitly casted to LED_mode type (notice, there isn't any need to cast LED_mode to int as it's done implicitly by compiler and no warning is generated).